Fixing basic CodeQL alerts (#384)
* Fixing basic CodeQL alerts

Many of these are just typing problems in debug logs and such. However,
more-defined behavior is better than undefined, so we'll center on
those. This may change some log lines but that's not really a surface
whose behaviors we plan to guarantee.

+++ libgearman-server/plugins/queue/redis/queue.cc
@@ -341,7 +341,7 @@ static gearmand_error_t _hiredis_add(gearman_server_st *, void *context,
   return gearmand_log_gerror(
     GEARMAN_DEFAULT_LOG_PARAM,
     GEARMAND_QUEUE_ERROR,
-    "failed to insert '%.*s' into redis", key.size(), &key[0]);
+    "failed to insert '%.*s' into redis", (uint32_t)key.size(), &key[0]);
 }
 
 static gearmand_error_t _hiredis_flush(gearman_server_st *, void *)
